+++ jardiniot-emb/arduino/src/include/motor/FanMotor.h
@@ -18,8 +18,8 @@ public:
     void reset();
 
     // Méthode hérité de la classe MotorStrategy
-    int activate();
-    int stop();
+    short activate();
+    short stop();
 };
 
 #endif
